fix jagged snapping behavior

+++ src/scalepicker.h
@@ -57,10 +57,10 @@ private:
     double firstPos; // converted to plot coordinates
     double firstPosPx; // pixel coordinates
     double currentPosPx; // current position in pixel coordinates
-    QList<double> snapPoints;
+    QList<int> snapPoints;
 
     double position(double); // returns the axis mouse position relative to plot coordinates
-    double positionPx(QMouseEvent*); // returns the axis mouse position in pixels
+    int positionPx(QMouseEvent*); // returns the axis mouse position in pixels
     double posCanvasPx(double pos); // returns the given position in canvas coordinates
 
 private slots:
